Move BboxAnnotationType to dicts.py to fix Pydantic TypeAdapter resolution
